# Warn if /var disk space is full and add a solution message