--- a/pkg/minikube/machine/start.go
+++ b/pkg/minikube/machine/start.go
@@ -151,6 +151,7 @@ func createHost(api libmachine.API, cfg *config.ClusterConfig, n *config.Node) (
 	if err != nil {
 		return nil, errors.Wrap(err, "new host")
 	}
+	defer postStartValidations(h, cfg.Driver)
 
 	h.HostOptions.AuthOptions.CertDir = localpath.MiniPath()
 	h.HostOptions.AuthOptions.StorePath = localpath.MiniPath()
@@ -205,24 +206,23 @@ func timedCreateHost(h *host.Host, api libmachine.API, t time.Duration) error {
 }
 
 // postStartValidations are validations against the host after it is created
-func postStartValidations(h *host.Host, drvName string) error {
+func postStartValidations(h *host.Host, drvName string) {
 	if !driver.IsKIC(drvName) {
-		return nil
+		return
 	}
 	// make sure /var isn't full, otherwise warn
 	output, err := h.RunSSHCommand("df -h /var | awk 'NR==2{print $5}'")
 	if err != nil {
-		return errors.Wrap(err, "df")
+		glog.Warning("error running df -h /var: %v", err)
 	}
 	output = strings.Trim(output, "\n")
 	percentageFull, err := strconv.Atoi(output[:len(output)-1])
 	if err != nil {
-		return err
+		glog.Warning("error getting % of /var that is free: %v", err)
 	}
 	if percentageFull >= 99 {
 		out.WarningT("The docker daemon is almost out of memory, run 'docker system prune' to free up space.")
 	}
-	return nil
 }
 
 // postStart are functions shared between startHost and fixHost
@@ -235,11 +235,6 @@ func postStartSetup(h *host.Host, mc config.ClusterConfig) error {
 
 	if driver.IsMock(h.DriverName) {
 		return nil
-	}
-
-	// first, run post start validations
-	if err := postStartValidations(h, mc.Driver); err != nil {
-		return errors.Wrap(err, "post-start validations")
 	}
 
 	glog.Infof("creating required directories: %v", requiredDirectories)