--- a/pkg/minikube/problem/err_map.go
+++ b/pkg/minikube/problem/err_map.go
@@ -582,7 +582,7 @@ var stateProblems = map[string]match{
 // dockerProblems are issues relating to issues with the docker driver
 var dockerProblems = map[string]match{
 	"NO_SPACE_ON_DEVICE": {
-		Regexp: re(`No space left on device`),
+		Regexp: re(`.*docker.*No space left on device.*`),
 		Advice: `Run 'docker system prune' to free up space on the device, or increase amount of memory allocated to Docker for Desktop via
 		
 	Docker icon > Settings > Resources > Disk Image Size