--- a/pkg/minikube/problem/err_map.go
+++ b/pkg/minikube/problem/err_map.go
@@ -583,9 +583,12 @@ var stateProblems = map[string]match{
 var dockerProblems = map[string]match{
 	"NO_SPACE_ON_DEVICE": {
 		Regexp: re(`.*docker.*No space left on device.*`),
-		Advice: `Run 'docker system prune' to free up space on the device, or increase amount of memory allocated to Docker for Desktop via
-		
-	Docker icon > Settings > Resources > Disk Image Size
+		Advice: `Try at least one of the following to free up space on the device:
+
+	1. Run "docker system prune" to remove unused docker data
+	2. Increase the amount of memory allocated to Docker for Desktop via 
+		Docker icon > Preferences > Resources > Disk Image Size
+	3. Run "minikube ssh -- docker system prune" if using the docker container runtime
 `,
 		Issues: []int{9024},
 	},