--- a/pkg/minikube/problem/problem.go
+++ b/pkg/minikube/problem/problem.go
@@ -104,6 +104,7 @@ func FromError(err error, goos string) *Problem {
 		netProblems,
 		deployProblems,
 		stateProblems,
+		dockerProblems,
 	}
 
 	var osMatch *Problem